Fix section table in database, add base operations for unit

+++ app/src/main/java/org/stepic/droid/model/Section.java
@@ -23,7 +23,7 @@ public class Section implements Serializable, Parcelable {
 
     private DateTimeFormatter mFormatForView;
 
-    private int id;
+    private long id;
     private long course; // course id
     private long[] units;
     private int position;
@@ -81,11 +81,11 @@ public class Section implements Serializable, Parcelable {
     }
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -249,7 +249,7 @@ public class Section implements Serializable, Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
+        dest.writeLong(this.id);
         dest.writeLong(this.course);
         dest.writeLongArray(this.units);
         dest.writeInt(this.position);
@@ -278,7 +278,7 @@ public class Section implements Serializable, Parcelable {
     }
 
     private Section(Parcel in) {
-        this.id = in.readInt();
+        this.id = in.readLong();
         this.course = in.readLong();
         this.units = in.createLongArray();
         this.position = in.readInt();
